// src/app/model/publication-data.ts
export class PublicationData {
    id: any;
    channel: string;
    publicationName: string;
    publicationType: string;
    contentType: string;
    subscriptionDate: any;
    expiryDate: any;
    language: string;
    category: string;
    country: string;
    region: string;
    benchMark: string;
    parrentPath: string;
    fileStrategy: string;
    regexForPage: string;
    regexForEdition: string;

}